feat: use NotificationProvider for push notification init
